Trait convert::cast::CastInto

pub trait CastInto<Dst, Neglect = ()> where
    Dst: CastFrom<Self, Neglect>,
    Neglect: UnsafeCastOptions, {
    fn cast_into(self) -> Dst
    where
        Self: Sized,
        Dst: Sized,
        Neglect: SafeCastOptions,
    { ... }

    unsafe fn unsafe_cast_into(self) -> Dst
    where
        Self: Sized,
        Dst: Sized,
        Neglect: UnsafeCastOptions,
    { ... }
}

🔬 This is a nightly-only experimental API. (cast)

Cast Self into Dst.

The reciprocal of CastFrom. This trait is implemented in terms of CastFrom.

Provided methods

fn cast_into(self) -> Dst where
Self: Sized,
Dst: Sized,
Neglect: SafeCastOptions,

🔬 This is a nightly-only experimental API. (cast)

Cast self into a value of type Dst, safely.

unsafe fn unsafe_cast_into(self) -> Dst where
Self: Sized,
Dst: Sized,
Neglect: UnsafeCastOptions,

🔬 This is a nightly-only experimental API. (cast)

Cast self into a value of type Dst, potentially unsafely.

Loading content...