Module convert::mem

🔬 This is a nightly-only experimental API. (cast)

(Extension) Traits for querying layout properties.

The definition of these traits demonstrate how TransmuteFrom can be used to query properties of a type's layout. See the Vec casting demonstration for an example of their use.

Traits

AlignEq	Experimental
Implemented if align_of::<Self>() == align_of::<Rhs>()

AlignLtEq	Experimental
Implemented if align_of::<Self>() <= align_of::<Rhs>()

SizeEq	Experimental
Implemented if size_of::<Self>() == size_of::<Rhs>()

SizeLtEq	Experimental
Implemented if size_of::<Self>() <= size_of::<Rhs>()