Module convert::transmute::options

Static checks that may be neglected when determining if two types are transmutable.

The default value of the Neglect parameter of TransmuteFrom and TransmuteInto, (), statically forbids transmutes that are unsafe, unsound, or unstable. However, you may explicitly opt-out of some static checks:

Transmute Option	Compromises	Usable With

NeglectStability	Stability	transmute_{from,into}, unsafe_transmute_{from,into}

NeglectAlignment	Safety	unsafe_transmute_{from,into}

NeglectValidity	Soundness	unsafe_transmute_{from,into}

The selection of multiple options is encoded by grouping them as a tuple; e.g., (NeglectAlignment, NeglectValidity) is a selection of both the NeglectAlignment and NeglectValidity options.

Structs

NeglectAlignment
Neglect the static alignment check.

NeglectStability
Neglect the static stability check.

NeglectValidity
Partially neglect the static validity check.

Traits

SafeTransmuteOptions
Options that may be used with safe transmutations.

UnsafeTransmuteOptions
Options that may be used with unsafe transmutations.