Trait convert::transmute::stability::PromiseTransmutableFrom

pub trait PromiseTransmutableFrom {
    type Archetype: TransmuteInto<Self, NeglectStability> + PromiseTransmutableFrom;
}

Promise that a type may be stably transmuted from other types.

To promise that all transmutations of any PromiseTransmutableInto type into your type that are currently safe will remain so in the future, simply annotate your type with #[derive(PromiseTransmutableFrom)].

For instance, this:

#[derive(PromiseTransmutableFrom)]
#[repr(C)]
pub struct Foo(pub Bar, pub Baz);

Associated Types

type Archetype: TransmuteInto<Self, NeglectStability> + PromiseTransmutableFrom

A type which exemplifies the greatest extent to which Self might change in non-breaking crate releases, insofar that those changes might affect instantiating Self via transmutation.

Loading content...