Trait convert::transmute::TransmuteInto

pub unsafe trait TransmuteInto<Dst: ?Sized, Neglect = ()> where
    Neglect: UnsafeTransmuteOptions, {
    fn transmute_into(self) -> Dst
    where
        Self: Sized,
        Dst: Sized,
        Neglect: SafeTransmuteOptions;

    unsafe fn unsafe_transmute_into(self) -> Dst
    where
        Self: Sized,
        Dst: Sized,
        Neglect: UnsafeTransmuteOptions;
}

Reinterpret the bits of Self as a type Dst.

The reciprocal of TransmuteFrom.

This trait is implemented automatically by the compiler for combinations of types where a transmutation is valid.

Required methods

fn transmute_into(self) -> Dst where
Self: Sized,
Dst: Sized,
Neglect: SafeTransmuteOptions,

Reinterpret the bits of a value of one type as another type, safely.

unsafe fn unsafe_transmute_into(self) -> Dst where
Self: Sized,
Dst: Sized,
Neglect: UnsafeTransmuteOptions,

Reinterpret the bits of a value of one type as another type, potentially unsafely.

The onus is on you to ensure that calling this method is safe.

Loading content...