Trait convert::transmute::stability::PromiseTransmutableInto

pub trait PromiseTransmutableInto {
    type Archetype: TransmuteFrom<Self, NeglectStability> + PromiseTransmutableInto;
}

Promise that a type may be stably transmuted into other types.

To promise that all safe transmutations from your type into other PromiseTransmutableFrom types will remain safe in the future, simply annotate your type with #[derive(PromiseTransmutableFrom)]; e.g.:

#[derive(PromiseTransmutableFrom)]
#[repr(C)]
pub struct Foo(pub Bar, pub Baz);

Associated Types

type Archetype: TransmuteFrom<Self, NeglectStability> + PromiseTransmutableInto

A type which exemplifies the greatest extent to which Self might change in non-breaking crate releases, insofar that those changes might affect converting Self into another type via transmutation.

Loading content...